+++ src/Http/Middleware/Middleware.php
@@ -5,17 +5,8 @@ namespace Rubix\Server\Http\Middleware;
 use Psr\Http\Message\ServerRequestInterface as Request;
 use Psr\Http\Message\ResponseInterface as Response;
 
-abstract class Middleware
+interface Middleware
 {
-    /**
-     * Run the middleware over the request.
-     *
-     * @param Request $request
-     * @param callable $next
-     * @return Response
-     */
-    abstract public function handle(Request $request, callable $next) : Response;
-
     /**
      * Allow an instance to be called like a function.
      *
@@ -23,8 +14,5 @@ abstract class Middleware
      * @param callable $next
      * @return Response
      */
-    public function __invoke(Request $request, callable $next) : Response
-    {
-        return $this->handle($request, $next);
-    }
+    public function __invoke(Request $request, callable $next) : Response;
 }